pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "otbtagreview"
version = "0.1.0"
description = "OTB Chess Tag Review - Convert video to PGN and Review"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Trae Builder", email = "builder@trae.ai" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "opencv-contrib-python>=4.7.0",
    "numpy>=1.24.0",
    "python-chess>=1.9.4",
    "click>=8.1.0",
]

[project.scripts]
otbtagreview = "otbtagreview.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["otbtagreview*"]
